Move code back to cbor.go to ease GitHub review
